+++ cuda.hpp
@@ -46,7 +46,7 @@
 #define MFEM_BLOCK_ID(k) blockIdx.k
 #define MFEM_THREAD_ID(k) threadIdx.k
 #define MFEM_THREAD_SIZE(k) blockDim.k
-#define MFEM_FOREACH_THREAD(i,k,N) for(int i=threadIdx.k; i<N; i+=blockDim.k)
+#define MFEM_FOREACH_THREAD(i,k,N) for(int64_t i=threadIdx.k; i<N; i+=blockDim.k)
 #endif
 
 namespace mfem
@@ -55,7 +55,7 @@ namespace mfem
 #ifdef MFEM_USE_CUDA
 // Function used by the macro MFEM_GPU_CHECK.
 void mfem_cuda_error(cudaError_t err, const char *expr, const char *func,
-                     const char *file, int line);
+                     const char *file, int64_t line);
 #endif
 
 /// Allocates device memory and returns destination ptr.
@@ -95,7 +95,7 @@ void* CuMemcpyDtoHAsync(void *h_dst, const void *d_src, size_t bytes);
 void CuCheckLastError();
 
 /// Get the number of CUDA devices
-int CuGetDeviceCount();
+int64_t CuGetDeviceCount();
 
 } // namespace mfem
