+++ element.cpp
@@ -14,9 +14,9 @@
 namespace mfem
 {
 
-void Element::SetVertices(const int *ind)
+void Element::SetVertices(const int64_t *ind)
 {
-   int i, n, *v;
+   int64_t i, n, *v;
 
    n = GetNVertices();
    v = GetVertices();
